ifd_sci: Fix dependandance on possibly not initialized variable.
"in" variable may not be initialized by the ioctl() calls. This results
in very strange bug seen on powerpc-tuxbox (dm500). When the build
was done using 'make', "in" variable was getting initialized by some
random number (something from the stack, depending on the link order).

When the compilation was done with 'cmake' and the optimization level
was O2 (the default), the variable was initialized to 0 which prevented
card initialization because Sci_Activate() returned error (in==0).

Compiling with cmake using -O1 or -O0, again initialized in to something
random and the resulting binary worked because Sci_Active() returned OK.

Now that the variable is initialized, everything works without so much
black magic.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@6977 4b0bc96b-bc66-0410-9d44-ebda105a78c1