Do not disable readers that return errors from ICC_Async_Get_Status().
Before moving PCSC to new card reader API only when sc8in1 reader was
used, ICC_Async_GetStatus() can ERROR. That in turn probably led to the fact
that the code in reader-coomn.c which should handle the ERROR to not
work very well.

Remove the offending code because it causing more harm than good.

This fixes ticket #2879.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@7773 4b0bc96b-bc66-0410-9d44-ebda105a78c1