csctapi: Make init procedure explicit.
It is not worthed to save 5 lines and have hard to follow code
in ICC_Async_Device_Init(). This makes it explicit what the init
procedure is for certain reader type.

For reference here are the reader types that matter to this code:
  #define R_DB2COM1   0x1 // Reader Dbox2 @ com1
  #define R_DB2COM2   0x2 // Reader Dbox2 @ com1
  #define R_SC8in1    0x3 // Reader Sc8in1 or MCR
  #define R_MOUSE     0x5 // Reader smartcard mouse

I have tested pure mouse (USB Phoenix) reader.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@7782 4b0bc96b-bc66-0410-9d44-ebda105a78c1