Added the chid calculation to all 0100 and 0500 caid's
I noticed that sid calculation is ussefull for all 0100 not only 6a,6c,6d
For 0500 it's also ussefull.

It's at base only give extra info about the ecm in log and thank's to this info it's possible to block ecm trough whitelist. Whithout it blockage trough oscam.whitelist does not work.

git-svn-id: https://svn.streamboard.tv/oscam/trunk@8706 4b0bc96b-bc66-0410-9d44-ebda105a78c1